Move Rect and Surface out of main file
