+++ SDLWrapper/begin_code.h
@@ -1,0 +1,5 @@
+namespace MiniEngine
+{
+
+inline namespace _SDLWrapper
+{
