+++ SDLWrapper/end_code.h
@@ -1,0 +1,3 @@
+} /// End of namespace MiniEngine::_SDLWrapper
+
+} /// End of namespace MiniEngine
