+++ SDLWrapper/include.h
@@ -1,0 +1,8 @@
+#pragma once
+
+/// Include SDL Library Headers.
+#include <SDL2/SDL.h>
+#undef main
+#include <SDL2/SDL_image.h>
+#include <SDL2/SDL_ttf.h>
+#include <SDL2/SDL_mixer.h>
